ai_analyzer/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.7
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.0
